+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/CombinedAttributeModel.kt
@@ -1,16 +1,17 @@
-package com.woocommerce.android.ui.products.variations.attributes
+package com.woocommerce.android.model
 
-import com.woocommerce.android.model.ProductAttribute
-import com.woocommerce.android.model.ProductGlobalAttribute
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
 
 /**
  * A "combined" product attribute is one that can be created from either a local (product-based) attribute
  * or a global (store-based) attribute
  */
+@Parcelize
 data class CombinedAttributeModel(
     val id: Long,
     val name: String
-) {
+) : Parcelable {
     companion object {
         fun fromLocalAttribute(attribute: ProductAttribute): CombinedAttributeModel {
             return CombinedAttributeModel(
@@ -27,6 +28,6 @@ data class CombinedAttributeModel(
         }
     }
 
-    val isGlobalAttribute: Boolean
-        get() = id != 0L
+    val isLocalAttribute: Boolean
+        get() = id == 0L
 }
