+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ProductNavigationTarget.kt
@@ -1,10 +1,11 @@
 package com.woocommerce.android.ui.products
 
+import com.woocommerce.android.model.CombinedAttributeModel
 import com.woocommerce.android.model.Product.Image
+import com.woocommerce.android.model.ProductFile
 import com.woocommerce.android.ui.products.ProductInventoryViewModel.InventoryData
 import com.woocommerce.android.ui.products.ProductPricingViewModel.PricingData
 import com.woocommerce.android.ui.products.ProductShippingViewModel.ShippingData
-import com.woocommerce.android.model.ProductFile
 import com.woocommerce.android.ui.products.settings.ProductCatalogVisibility
 import com.woocommerce.android.ui.products.settings.ProductVisibility
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event
@@ -75,4 +76,7 @@ sealed class ProductNavigationTarget : Event() {
         ProductNavigationTarget()
     object AddProductDownloadableFile : ProductNavigationTarget()
     object AddProductAttribute : ProductNavigationTarget()
+    data class AddProductAttributeTerms(
+        val combinedAttribute: CombinedAttributeModel
+    ): ProductNavigationTarget()
 }
